# core-common-kycb-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-kycb-mgmt
  description: KYC/B & AML Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/kycb
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/corporate/parties/{partyId}/structure:
    get:
      tags:
        - Corporate Structure
      summary: List structure relationships
      description: Retrieves all corporate structure relationships with filtering capabilities
      operationId: listStructureRelationships
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.partyId
          in: query
          description: Exact filter for party id
          required: false
          schema:
            type: string
        - name: filters.parentEntityId
          in: query
          description: Exact filter for parent entity id
          required: false
          schema:
            type: string
        - name: filters.ownershipPercentage
          in: query
          description: Exact filter for ownership percentage
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[ownershipPercentage].from
          in: query
          description: Filter ownership percentage from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[ownershipPercentage].to
          in: query
          description: Filter ownership percentage to value
          required: false
          schema:
            type: string
        - name: filters.relationshipType
          in: query
          description: Exact filter for relationship type
          required: false
          schema:
            type: string
        - name: filters.controlNotes
          in: query
          description: Exact filter for control notes
          required: false
          schema:
            type: string
        - name: filters.isVerified
          in: query
          description: Exact filter for is verified
          required: false
          schema:
            type: boolean
        - name: filters.verificationDate
          in: query
          description: Exact filter for verification date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[verificationDate].from
          in: query
          description: Filter verification date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[verificationDate].to
          in: query
          description: Filter verification date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.startDate
          in: query
          description: Exact filter for start date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[startDate].from
          in: query
          description: Filter start date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[startDate].to
          in: query
          description: Filter start date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.endDate
          in: query
          description: Exact filter for end date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[endDate].from
          in: query
          description: Filter end date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[endDate].to
          in: query
          description: Filter end date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved corporate structure relationships
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
    post:
      tags:
        - Corporate Structure
      summary: Add structure relationship
      description: Adds a new structure relationship to the specified party
      operationId: addStructureRelationship
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorporateStructureDTO'
        required: true
      responses:
        '201':
          description: Successfully added structure relationship
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorporateStructureDTO'
...
  /api/v1/documents/corporate/{documentId}:
    get:
      tags:
        - Corporate Documents
      summary: Get specific corporate document
      description: Retrieves a specific corporate document by its ID
      operationId: getCorporateDocument
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved corporate document
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorporateDocumentDTO'
        '404':
          description: Corporate document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorporateDocumentDTO'
    delete:
      tags:
        - Corporate Documents
      summary: Delete corporate document
      description: Deletes a corporate document
      operationId: deleteCorporateDocument
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted corporate document
        '404':
          description: Corporate document not found
    patch:
      tags:
        - Corporate Documents
      summary: Update corporate document
      description: Updates an existing corporate document
      operationId: updateCorporateDocument
      parameters:
        - name: documentId
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorporateDocumentDTO'
        required: true
      responses:
        '200':
          description: Successfully updated corporate document
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorporateDocumentDTO'
        '404':
          description: Corporate document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorporateDocumentDTO'
...
  /api/v1/corporate/parties/{partyId}/structure/{structureId}:
    get:
      tags:
        - Corporate Structure
      summary: Get specific structure relationship
      description: Retrieves a specific structure relationship by its ID
      operationId: getStructureRelationship
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: structureId
          in: path
          description: ID of the structure relationship
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved structure relationship
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorporateStructureDTO'
        '404':
          description: Structure relationship not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorporateStructureDTO'
    delete:
      tags:
        - Corporate Structure
      summary: Delete structure relationship
      description: Deletes a structure relationship
      operationId: deleteStructureRelationship
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: structureId
          in: path
          description: ID of the structure relationship
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted structure relationship
        '404':
          description: Structure relationship not found
    patch:
      tags:
        - Corporate Structure
      summary: Update structure relationship
      description: Updates an existing structure relationship
      operationId: updateStructureRelationship
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: structureId
          in: path
          description: ID of the structure relationship
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorporateStructureDTO'
        required: true
      responses:
        '200':
          description: Successfully updated structure relationship
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorporateStructureDTO'
        '404':
          description: Structure relationship not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CorporateStructureDTO'
...
components:
  schemas:
...
    CorporateDocumentDTO:
      required:
        - documentReference
        - documentType
        - isVerified
        - issueDate
        - partyId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        corporateDocumentId:
          type: string
          format: uuid
          readOnly: true
        partyId:
          type: string
          format: uuid
        documentType:
          maxLength: 50
          minLength: 0
          type: string
        documentReference:
          maxLength: 100
          minLength: 0
          type: string
        documentSystemId:
          maxLength: 100
          minLength: 0
          type: string
        notaryName:
          maxLength: 200
          minLength: 0
          type: string
        notaryLocation:
          maxLength: 200
          minLength: 0
          type: string
        commercialRegistry:
          maxLength: 200
          minLength: 0
          type: string
        registryEntry:
          maxLength: 100
          minLength: 0
          type: string
        issueDate:
          type: string
          format: date-time
        registryDate:
          type: string
          format: date-time
        expiryDate:
          type: string
          format: date-time
        isVerified:
          type: boolean
        verificationNotes:
          maxLength: 1000
          minLength: 0
          type: string
        verificationDate:
          type: string
          format: date-time
        verificationAgent:
          maxLength: 100
          minLength: 0
          type: string
...
    CorporateStructureDTO:
      required:
        - isVerified
        - ownershipPercentage
        - parentEntityId
        - partyId
        - relationshipType
        - startDate
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        corporateStructureId:
          type: string
          format: uuid
          readOnly: true
        partyId:
          type: string
          format: uuid
        parentEntityId:
          type: string
          format: uuid
        ownershipPercentage:
          type: number
        relationshipType:
          maxLength: 50
          minLength: 0
          type: string
        controlNotes:
          maxLength: 1000
          minLength: 0
          type: string
        isVerified:
          type: boolean
        verificationDate:
          type: string
          format: date-time
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.